--- waypoint_patroller/launch/patroller.launch
@@ -1,10 +1,0 @@
-<launch>
-
-    <arg name="waypoints"/>
-    <arg name="randomized" default="true"/>
-
-    
-
-
-    <node name="patroller" pkg="waypoint_patroller" type="patroller.py" output="screen" args="$(arg waypoints) $(arg randomized)"/>
-</launch>
