--- waypoint_patroller/launch/talking_patroller.launch
@@ -1,10 +1,0 @@
-<launch>
-
-
-    <node name="patroller" pkg="patroller" type="talking_patroller.py">
-	 <param name="waypoints" value="$(find waypoint_recorder)/waypoints/waypoints.csv" />
-    </node>	
-
-
-
-</launch>
