+++ waypoint_patroller/scripts/long_term_patroller.py
@@ -1,155 +1,80 @@
 #! /usr/bin/env python
+
 import sys
-from random import shuffle
 import rospy
 import csv
 
 import smach
 import smach_ros
 
-
-from navigation import navigation
-from charging import dock_and_charge
-
-from scitos_msgs.msg import BatteryState
-
+from waypoint_patroller.patroller import WaypointPatroller
 
 from dynamic_reconfigure.server import Server
 from waypoint_patroller.cfg import BatteryTresholdsConfig
-from geometry_msgs.msg import Pose
 
-import strands_datacentre as datacentre
 import strands_datacentre.util
 got_pymongo = strands_datacentre.util.check_for_pymongo()
 if got_pymongo:
     import pymongo
-
-# ParameterStore is a singleton class that contains all the battery
-# tresholds. It is used so we that the updated tresholds can then be read
-# by the battery monitor in monitor_states.py
+    
 from parameter_store import ParameterStore
 
-# This file implements the higher level state machine for long term
-# patrolling. It uses both the navigation and the dock and charge state
-# machines
-
-
-# reconfigures the battery tresholds
-def reconfigure_callback(config, level):
-    ParameterStore().CHARGED_BATTERY = config.charged_battery
-    ParameterStore().LOW_BATTERY = config.low_battery
-    ParameterStore().VERY_LOW_BATTERY = config.very_low_battery
-    return config
-
-
-# the point chooser state checks the battery life, and if it is greater
-# than CHARGE_BATTERY_TRESHOLD, sends the robot to a new patrol point.
-# Otherwise, the robot is sent to the charging station (assumed to be the
-# first point in the waypoints file). The ordering of visitng the
-# patrolling points is either sequential or random, depending on a
-# command-line argument, as is the number of iterations the robot should
-# do until the state machine terminates with success
-class PointChooser(smach.State):
-
+class LongTermPatroller(object):
+    """
+    Constructor.
+    :param waypoints_name: str, name of waypoints as appears in the datacentre
+    :param is_random: bool, should the waypoints be visited in random order
+    :param n_iterations: int, how many times to visit all the waypoints
+    """
     def __init__(self, waypoints_name, is_random, n_iterations):
-        smach.State.__init__(self,
-                             outcomes=[
-                                 'patrol', 'go_charge', 'succeeded'],
-                             output_keys=['goal_pose', 'going_to_charge']
-                             )
+    
+        # Create the main state machine
+        self.long_term_patrol_sm = WaypointPatroller(waypoints_name, is_random,
+                                                n_iterations)
+        
+        # dynamic reconfiguration of battery tresholds
+        self.srv = Server(BatteryTresholdsConfig, self.reconfigure_callback)
+    
+        pass
+        
+    """ Dyanmic reconfigure callback for the battery tresholds """
+    def reconfigure_callback(self, config, level):
+        ParameterStore().CHARGED_BATTERY = config.charged_battery
+        ParameterStore().LOW_BATTERY = config.low_battery
+        ParameterStore().VERY_LOW_BATTERY = config.very_low_battery
+        return config
 
-        self.points = self._get_points(waypoints_name)  # []
-        self.point_set = waypoints_name
-        self.charing_station_pos = [waypoints_name, "charging_point"]
-
-        self.current_point = -1
-        self.n_points = len(self.points)
-
-        self.is_random = is_random
-        self.n_iterations = n_iterations
-        self.iterations_completed = 0
-
-        # rearranges the list of points to visit randomly
-        if self.is_random:
-            shuffle(self.points)
-
-        self.battery_life = 100
-        self.battery_monitor = rospy.Subscriber(
-            "/battery_state", BatteryState, self.bat_cb)
-
-    """ Get a list of points in the given set """
-    def _get_points(self, point_set):
-        mongo = pymongo.MongoClient(rospy.get_param("datacentre_host"),
-                                    rospy.get_param("datacentre_port"))
-        points = []
-        search =  {"meta.pointset": point_set}
-        for point in mongo.autonomous_patrolling.waypoints.find(search):
-            print point
-            if point["meta"]["name"] != "charging_point":
-                points.append([point_set, point["meta"]["name"]])
-        return points
-
-    """ Get a given waypoint pose """
-    def _get_point(self, point_name, point_set):
-        mongo = pymongo.MongoClient(rospy.get_param("datacentre_host"),
-                                    rospy.get_param("datacentre_port"))
-        search = {"meta.name": point_name,
-                  "meta.pointset": point_set}
-        p = mongo.autonomous_patrolling.waypoints.find(search)            
-        p = p[0]
-        meta, pose = strands_datacentre.util.document_to_msg(p, Pose)
-        return pose
+    
+    """ The Main start point for Long Term Patroller """
+    def main(self):
+        # Execute SMACH plan
+        sis = smach_ros.IntrospectionServer('server_name',
+                                            self.long_term_patrol_sm,
+                                            '/SM_ROOT')
+        sis.start()
+        outcome = self.long_term_patrol_sm.execute()
+    
+        rospy.spin()
+        sis.stop()
 
 
-    def bat_cb(self, msg):
-        self.battery_life = msg.lifePercent
-
-    def execute(self, userdata):
-        rospy.sleep(1)
-
-        if self.battery_life > ParameterStore().LOW_BATTERY + 5:
-            self.current_point = self.current_point + 1
-            if self.current_point == self.n_points:
-                self.iterations_completed = self.iterations_completed + 1
-                if self.iterations_completed == self.n_iterations:
-                    return 'succeeded'
-                self.current_point = 0
-                if self.is_random:
-                    shuffle(self.points)
-
-            current_pt = self.points[self.current_point]
-            userdata.goal_pose = self._get_point(
-                current_pt[1], current_pt[0])  # current_row
-            userdata.going_to_charge = 0
-            return 'patrol'
-        else:
-            userdata.goal_pose = self._get_point("charging_point",
-                                                 self.point_set)
-            userdata.going_to_charge = 1
-            return 'go_charge'
-
-
-def main():
-
+if __name__ == '__main__':
     rospy.init_node('patroller')
-
+    
     # Check for connection to the datacentre
     got_mongo = strands_datacentre.util.wait_for_mongo()
-    if not got_mongo:
-        sys.exit(1)
     if not got_pymongo:
         sys.exit(1)
-
-    # dynamic reconfiguration of battery tresholds
-    srv = Server(BatteryTresholdsConfig, reconfigure_callback)
+    if not got_mongo:
+        sys.exit(1)
 
     # Check if a waypoints file was given as argument
     if len(sys.argv) < 2:
-        rospy.logerr("No waypoints file given. Use rosrun waypoint_patroller " 
-                     "patroller.py [path to csv waypoints file]. If you are " 
-                     "using a launch file, see launch/patroller.launch for " 
+        rospy.logerr("No waypoints dataset name given. Use rosrun " 
+                     "waypoint_patroller patroller.py [dataset name]. If you " 
+                     "are using a launch file, see launch/patroller.launch for " 
                      "an example.")
-        return 1
+        sys.exit(1)
 
     # waypoints file is a csv file with goal poses. The first line of the file
     # has the position in front of the charging station
@@ -180,44 +105,6 @@ def main():
     else:
         rospy.loginfo("Executing waypoint_patroller with infinite iterations")
 
-    # Create a SMACH state machine
-    long_term_patrol_sm = smach.StateMachine(outcomes=['succeeded', 'aborted'])
-    with long_term_patrol_sm:
-        smach.StateMachine.add('POINT_CHOOSER',
-                               PointChooser(waypoints_name,
-                                            is_random,
-                                            n_iterations),
-                               transitions={'patrol': 'PATROL_POINT',
-                                            'go_charge': 'GO_TO_CHARGING_STATION',
-                                            'succeeded': 'succeeded'})
-        smach.StateMachine.add('PATROL_POINT',
-                               navigation(),
-                               transitions={'succeeded': 'POINT_CHOOSER',
-                                            'battery_low': 'POINT_CHOOSER',
-                                            'bumper_failure': 'aborted',
-                                            'move_base_failure': 'POINT_CHOOSER'})
-        smach.StateMachine.add('GO_TO_CHARGING_STATION',
-                               navigation(),
-                               transitions={'succeeded': 'DOCK_AND_CHARGE',
-                                            'battery_low': 'GO_TO_CHARGING_STATION',
-                                            'bumper_failure': 'aborted',
-                                            'move_base_failure': 'aborted'})
         
-        smach.StateMachine.add('DOCK_AND_CHARGE',
-                               dock_and_charge(),
-                               transitions={'succeeded': 'POINT_CHOOSER',
-                                            'failure': 'aborted'})
-
-    # Execute SMACH plan
-    sis = smach_ros.IntrospectionServer('server_name',
-                                        long_term_patrol_sm,
-                                        '/SM_ROOT')
-    sis.start()
-    outcome = long_term_patrol_sm.execute()
-
-    rospy.spin()
-    sis.stop()
-
-
-if __name__ == '__main__':
-    main()
+    l =  LongTermPatroller(waypoints_name, is_random, n_iterations)
+    l.main()
