+++ waypoint_patroller/src/waypoint_patroller/__init__.py
@@ -1,0 +1,8 @@
+import charging
+import monitor_states
+import navigation
+import recover_states
+import patroller
+
+__all__ = ['navigation', 'recover_states', 'monitor_states', 'charging',
+           'patroller']
